refactor: 🎨 refactor actions
